% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{acabamento}
\alias{acabamento}
\title{Acabamento de superfície de peças metálicas}
\format{
Data frame com 18 linhas e três variáveis:
\itemize{
  \item acabamento: variável resposta
  \item tinta: tipo de tinta utilzada (fator 1; dois níveis)
  \item tempo: tempo de secagem (fator 2; 3 níveis: 20min, 25min e 30min)
}

Data frame com 20 linhas e três variáveis:
\itemize{
  \item corrente: variável resposta
  \item polissilicone: tratamento com polissiliconte (fator 1; dois níveis - 1e-20; 2e-20)
  \item recozimento: recozimento (fator 2; 5 níveis: 900/60, 900/180, 950/60, 1000/15, 1000/30 - temperatura/tempo)
}
}
\description{
Exercício 14.2: Um engenheiro suspeita que o acabamento de uma superfícies de peças metálicas seja influenciado pelo tipo de tinta usada e pelo tempo de secagem. Ele selecionou 3 tempos de secagem - 20, 25, e 30 minutos - e usou dois tipos de tinta. Três peças são testadas com cada combinação de tipo de tinta e tempo de secagem. Realize a análise dos dados e tire as suas conclusões.

Exercício 14.7: Um engenheiro realizou um estudo sobre o efeito de duas variáveis -  tratamento com polissilicone e condições de recozimento (tempo e temperatura) - sobre a corrente de um transmissor biploar. Analise os dados provenientes desse experimento e tire as suas conclusões.
}
\references{
Montgomery, D. C., Runger, G. C. (2015) Estatística Aplicada e Probabilidade para Engenheiros, 5ª. Edição, Rio de Janeiro: LTC.

Montgomery, D. C., Runger, G. C. (2015) Estatística Aplicada e Probabilidade para Engenheiros, 5ª. Edição, Rio de Janeiro: LTC.
}
\author{
Fábio N. Demarqui \email{fndemarqui@est.ufmg.br}

Fábio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{datasets}
