% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{algodao}
\alias{algodao}
\title{Nível de colesterol no sangue}
\format{
Data frame com 25 linhas e duas variáveis:
\itemize{
  \item percentual: percentual de algodão utilizado na mistura do material da fibra
  \item resistencia: resistência da fibra sintética (variável resposta).
}
}
\description{
Exercício 3.10: Um engenheiro está estudando se a resistência de um novo tipo de fibra sintética pode ser usada para a confecção de camisetas masculinas. Sabendo que a resistência é, em geral, afetada pelo percentual de algodão utilizado na mistura do material da fibra, o engenheiro decidiu conduzir um experimento com 5 níveis (percentuais) de algodão e 5 repetições para cada nível.
}
\references{
Montgomery, D. C. (2016). Design and Analysis of Experiments, 8ª. Edição, New York: Wiley.
}
\author{
Fábio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{datasets}
