% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{alimentacao}
\alias{alimentacao}
\title{Programa de modificação alimentar}
\format{
Data frame com 20 linhas e duas variáveis:
\itemize{
\item medida: quando o peso do participante foi medido
\item peso: peso do participante
}
}
\description{
Exercício 10.43: Dez indivíduos participaram de um programa de modificação alimentar para estimular a perda de peso. Os pesos dos participantes foram registrados antes e depois da participação no programa.
}
\references{
Montgomery, D. C., Runger, G. C. (2015) Estatística Aplicada e Probabilidade para Engenheiros, 5ª. Edição, Rio de Janeiro: LTC.
}
\author{
Fábio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{datasets}
