% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{aluminio}
\alias{aluminio}
\title{Resistência à tensão no alumínio}
\format{
Data frame com 22 linhas e duas variáveis:
\itemize{
  \item resistencia: resistência à tensão da estrutura de alumínio
  \item estrutura: tipo de estrutura (E1 ou E2)
}
}
\description{
Exemplo 10.4 (adaptado): Testes d resistência à tensão foram feitos em dois tipos diferentes de estruturas de alumínio. Essas estruturas foram usadas na fabricação das asas de um avião comercial. De experiências passadas com o processo de fabricação dessas estruturas e com o procedimento de testes, os desvios-padrão são considerados conhecidos (\eqn{\sigma_{1} = 1} e \eqn{\sigma_{2} = 1.5}, respectivamente).
}
\references{
Montgomery, D. C., Runger, G. C. (2015) Estatística Aplicada e Probabilidade para Engenheiros, 5ª. Edição, Rio de Janeiro: LTC.
}
\author{
Fábio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{datasets}
