% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{arsenio}
\alias{arsenio}
\title{Arsênio em agua potável}
\format{
Data frame com 20 linhas e duas variáveis:
\itemize{
\item comunidade: tipo de comunidade onde a amostra de água foi coletada (metropolitana ou rural)
\item concentracao: concentração de arsênio presente na amostra (variável resposta)
}
}
\description{
Exemplo 10.6: A concentração de arsênio em suprimentos públicos de água potável é um reisco potencial de saúde. Um estudo reportou a concentração, em partes por bilhão (ppb), de arsênio em água potável para 10 comunidades metropolitanas de Fênix e 10 comunidades rurais do Arizona.
}
\references{
Montgomery, D. C., Runger, G. C. (2015) Estatística Aplicada e Probabilidade para Engenheiros, 5ª. Edição, Rio de Janeiro: LTC.
}
\author{
Fábio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{datasets}
