% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biblioteca.R
\name{as.factor2k}
\alias{as.factor2k}
\title{Canonical codification for $\eqn{2^k}$ fatoctorial designs}
\usage{
as.factor2k(x)
}
\arguments{
\item{x}{factor to be converted.}
}
\value{
the recodified factor with -1 and +1 levels
}
\description{
Function to convert the low and hight levels of a given factor to -1 and 1, respectively.
}
\examples{
library(planex)
library(tidyverse)
data(rugosidade)
head(rugosidade)
rugosidade <- mutate(rugosidade,
  alimentacao = as.factor2k(alimentacao),
  profundidade = as.factor2k(profundidade),
  angulo = as.factor2k(angulo)
)
rugosidade


}
