% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{bocais}
\alias{bocais}
\title{Efeitos do tipo de bocal sobre medidas de formas}
\format{Data frame com 30 linhas e três variáveis:
\itemize{
  \item medida: medida da forma (variável resposta)
  \item bocal: tipo de bocal (variável categórica com 5 níveis; fator de interesse)
  \item velocidade: velocidade do jato de saída, em m/s; (bloco; fator de interesse secundário)
}}
\description{
(Exercício 13.37) Delineamento em esquema de blocos completos aleatorizados e um fator fixo conduzido para avaliar o efeito de 5 diferentes tipos de bocais sobre sobre a medida de forma.
}
\references{
Montgomery, D. C., Runger, G. C. (2015) Estatística Aplicada e Probabilidade para Engenheiros, 5ª. Edição, Rio de Janeiro: LTC.
}
\author{
Fabio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{datasets}
