% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{chocolate}
\alias{chocolate}
\title{Consumo de chocolate sobre saúde cardiovascular}
\format{Data frame com 36 linhas e duas variáveis:
\itemize{
  \item capacidade: capacidade antioxidante do plasma sanguíneo (variável resposta)
  \item tipo: tipo de chocolate consumido (variável categórica com quadro níveis; fator de interesse)
}}
\description{
Delineamento com um fator fixo conduzido para investigar o efeito do consumo de chocolate sobre saúde cardiovascular. O experimento consistiu no uso de três tipos diferentes de chocolates: 100 g de chocolate, 100 g de chocolate amargo com 200 mL de leite integral, e 200 g de chocolate ao leite. Os voluntários participantes do estudo receberam aleatoriamente um dos 3 tipos de chocolates investigados e depois de uma hora do consumo do produto a capacidade antioxidante do plasma sanguíneo foi medida em um ensaio.
}
\references{
Montgomery, D. C. (2016) Design and Analysis of Experiments, 8ª. Edição, New York: Wiley.
}
\author{
Fabio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{datasets}
