% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{chumbo}
\alias{chumbo}
\title{Acabamento de superfície de peças metálicas}
\format{
Data frame com 16 linhas e quatro variáveis:
\itemize{
  \item recuperacao: recuperação de chumbo (%) - variável resposta
  \item TA: tempo de agitação (fator 1 - nível baixo: 5e-6; nível alto: 5e-5)
  \item pH: pH (fator 2 - nível baixo: 6; nível alto: 8)
  \item CR: concentração de reagente (fator 3 - nível baixo: 10; nível alto: 30)
}
}
\description{
Exercício 14.26: Um delineamento 2^3 foi conduzido para determinar o nível de chumbo, utilizando-se a espectrometria de absorção atômica de chama (EAAC). Análise adequadamente os dados provenientes desse experimento, apresentando as suas conclusões.
}
\references{
Montgomery, D. C., Runger, G. C. (2015) Estatística Aplicada e Probabilidade para Engenheiros, 5ª. Edição, Rio de Janeiro: LTC.
}
\author{
Fábio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{datasets}
