% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{colesterol}
\alias{colesterol}
\title{Nível de colesterol no sangue}
\format{
Data frame com 30 linhas e duas variáveis:
\itemize{
\item medida: quando o peso do participante foi medido
\item nivel: nível de colesterol no sangue
}
}
\description{
Exercício 10.41: Quinze homens adultos, entre as idades de 35 e 50 anos, participaram de um estudo para avaliar o efeito da dieta e de exercícios no nível de colesterol no sangue. O colesterol total foi medido para cada indivíduo inicialmente e depois de 3 meses de participação em um programa de exercícios aeróbicos e de mudanças para uma dieta de baixo teor de gordura.
}
\references{
Montgomery, D. C., Runger, G. C. (2015) Estatística Aplicada e Probabilidade para Engenheiros, 5ª. Edição, Rio de Janeiro: LTC.
}
\author{
Fábio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{datasets}
