% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{corrente}
\alias{corrente}
\title{Acabamento de superfície de peças metálicas}
\format{
Data frame com 20 linhas e três variáveis:
\itemize{
\item corrente: variável resposta
\item polissilicone: tratamento com polissiliconte (fator 1; dois níveis - 1e-20; 2e-20)
\item recozimento: recozimento (fator 2; 5 níveis: 900/60, 900/180, 950/60, 1000/15, 1000/30 - temperatura/tempo)
}
}
\description{
Exercício 14.7: Um engenheiro realizou um estudo sobre o efeito de duas variáveis -  tratamento com polissilicone e condições de recozimento (tempo e temperatura) - sobre a corrente de um transmissor biploar. Analise os dados provenientes desse experimento e tire as suas conclusões.
}
\references{
Montgomery, D. C., Runger, G. C. (2015) Estatística Aplicada e Probabilidade para Engenheiros, 5ª. Edição, Rio de Janeiro: LTC.
}
\author{
Fábio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{datasets}
