% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{densidade}
\alias{densidade}
\title{Densidade cozida de um ânodo de carbono.}
\format{
Data frame com 18 linhas e 3 variáveis:
\itemize{
\item posicao: posição do forno - fator fixo com 2 níveis
\item temperatura: temperatura de queima (°C) - fator fixo com 3 níveis
\item densidade: densidade cozida do ânodo de carbono (variável resposta)
}
}
\description{
Um experimento foi conduzido para determinar se a temperatura de queima ou a posição do forno afetam a densidade cozida de um ânodo de carbono.
}
\references{
Montgomery, D. C. (2016). Design and Analysis of Experiments, 8ª. Edição, New York: Wiley.
}
\author{
Fábio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{datasets}
