% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{dietas}
\alias{dietas}
\title{Efeito de dietas sobre o teor de proteínas do leite}
\format{Data frame com 79 linhas e duas variáveis:
\itemize{
  \item teor: teor de proteína do leite (variável resposta)
  \item dieta: tipo de dieta (variável categórica com três níveis; fator de interesse)
}}
\description{
(Exercício 13.13) Delineamento com um fator fixo conduzido para investigar o efeito de três dietas sobre o teor de proteína de leite de vaca.
}
\references{
Montgomery, D. C., Runger, G. C. (2015) Estatística Aplicada e Probabilidade para Engenheiros, 5ª. Edição, Rio de Janeiro: LTC.
}
\author{
Fabio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{datasets}
