% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{enxertos}
\alias{enxertos}
\title{Enxertos vasculares (veias artificiais)}
\format{
Data frame com 24 linhas e 3 variáveis:
\itemize{
\item pressao: pressão de extrusão (em libras; fator de interesse principal)
\item lote: lote de materia prima (bloco - fator de interesse secundário)
\item saliencia: saliência (variável resposta)
}
}
\description{
Um delineamento com blocos completos aleatorizados foi realizado com o objetivo de avaliar o efeito de diferentes níveis de pressão na ocorrência de saliências em enxertos vasculares produzidos com diferentes lotes de materia prima (fonte de variablidade de perturbação). Analise adequadamente os dados provenientes desse experimento, apresentando as conclusões práticas.
}
\references{
Montgomery, D. C. (2016). Design and Analysis of Experiments, 8ª. Edição, New York: Wiley.
}
\author{
Fábio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{datasets}
