% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{escapamento}
\alias{escapamento}
\title{Efeitos do tipo de bocal sobre medidas de formas}
\format{Data frame com 25 linhas e três variáveis:
\itemize{
  \item escapamento: escapamento de corrente elétrica (variável resposta)
  \item comprimento: comprimento do canal (variável categórica com 4 níveis; fator de interesse)
  \item largura: largura do canal (bloco; fator de interesse secundário)
}}
\description{
(Exercício 13.41) Delineamento em esquema de blocos completos aleatorizados e um fator fixo conduzido para avaliar o efeito de 4 comprimentos de canais sobre o escapamento de corrente elétrica.
}
\references{
Montgomery, D. C., Runger, G. C. (2015) Estatística Aplicada e Probabilidade para Engenheiros, 5ª. Edição, Rio de Janeiro: LTC.
}
\author{
Fabio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{datasets}
