% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{filtragem}
\alias{filtragem}
\title{Taxa de filtragem de um produto químico}
\format{
Data frame com 16 linhas e 5 variáveis:
\itemize{
  \item temperatura: taxa de agitação (variável categórica com 2 níveis; fator4)
  \item pressao: taxa de agitação (variável categórica com 2 níveis; fator4)
  \item concentracao: taxa de agitação (variável categórica com 2 níveis; fator4)
  \item agitacao: taxa de agitação (variável categórica com 2 níveis; fator4)
  \item filtragem: taxa de filtragem (variável resposta)
  \item comb_trat: combinação de tratamentos (ausência da letra indica nível baixo; presença, nível alto)
}
}
\description{
Exemplo 6.2: Um produto químico é produzido em um vaso de pressão. Um experimento fatorial é realizado na planta piloto para estudar os fatores que podem vir a influenciar a taxa de filtragem desse produto. Os quatro fatores são temperatura (A), pressão (B), concentração de formaldeído (C) e taxa de agitação (D). Cada fator está presente em dois níveis. O desenho da matriz e os dados de resposta obtidos de uma única réplica. As 16 corridas são feitas em ordem aleatória. O engenheiro de processos está interessado em maximizar a taxa de filtração.
}
\references{
Montgomery, D. C. (2016) Design and Analysis of Experiments, 8ª. Edição, New York: Wiley.
}
\author{
Fábio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{datasets}
