% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{filtragemFrac}
\alias{filtragemFrac}
\title{Taxa de filtragem de um produto químico (meia fração)}
\format{
Data frame com 8 linhas e 2 variáveis:
\itemize{
\item filtragem: taxa de filtragem (variável resposta)
\item comb_trat: combinação de tratamentos (ausência da letra indica nível baixo; presença, nível alto)
}
}
\description{
Exercício 14.44: Um produto químico é produzido em um vaso de pressão. Um experimento fatorial fracionário (meia fração) é realizado na planta piloto para estudar os fatores que podem vir a influenciar a taxa de filtragem desse produto. Os quatro fatores são temperatura (A), pressão (B), concentração de formaldeído (C) e taxa de agitação (D). Cada fator está presente em dois níveis. . O engenheiro de processos está interessado em maximizar a taxa de filtração.
}
\references{
Montgomery, D. C., Runger, G. C. (2015) Estatística Aplicada e Probabilidade para Engenheiros, 5ª. Edição, Rio de Janeiro: LTC.
}
\author{
Fábio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{datasets}
