% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{fogo}
\alias{fogo}
\title{Tratamentos de retardantes de fogo}
\format{
Data frame com 16 linhas e duas variáveis:
\itemize{
  \item polegadas: polegadas de tecido queimado na amostra de teste de tamamnho padrão
  \item combtrat: combinação de tratamentos
}
}
\description{
Exercício 14.18: Um engenheiro conduziu um experimento para testar tecidos com relação à chama, depois de aplicar tratamentos com retardantes de fogo. Os quatro fatores considerados são: o tipo de tecido (A), o tipo de tratamento de retardante de fogo (B), condição de lavagem (C -  o nível baixo é sem lavagem; o nível alto é depois de uma lavagem), e o método de condução do teste de chama (D). Todos os fatores foram corridos em dois níveis e a variável resposta são as polegadas de tecido queimado na amostra de teste de tamanho padrão. Análise adequadamente os dados provenientes desse experimento, apresentando as suas conclusões.
}
\references{
Montgomery, D. C., Runger, G. C. (2015) Estatística Aplicada e Probabilidade para Engenheiros, 5ª. Edição, Rio de Janeiro: LTC.
}
\author{
Fábio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{datasets}
