% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{forca}
\alias{forca}
\title{Força de fibras sintéticas}
\format{
Data frame com 24 linhas e 3 variáveis:
\itemize{
  \item forca: variável resposta.
  \item operador: fator aleatório com 3 níveis.
  \item maquina: fator fixo com 4 níveis.
}
}
\description{
Exercício 13.3: Os fatores que influenciam a resistência à ruptura de um fibras sintéticas estão sendo estudados. As quatro máquinas de produção disponíveis e três operadores são escolhidos e um experimento fatorial é executado usando fibras de um mesmo lote de produção.
}
\references{
Montgomery, D. C. (2016) Design and Analysis of Experiments, 8ª. Edição, New York: Wiley.
}
\author{
Fábio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{datasets}
