% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biblioteca.R
\name{Fpower}
\alias{Fpower}
\title{Power function for F tests involving one-way and two-way ANOVA}
\usage{
Fpower(alpha = 0.05, nlev = NULL, nreps = NULL, Delta = NULL, sigma = NULL)
}
\arguments{
\item{alpha}{the significance level of the test (default value is alpha = 0.05).}

\item{nlev}{vector containing the number of levels of the factors.}

\item{nreps}{the number of replicates in each combination of factor levels.}

\item{Delta}{the size of a practical difference in two means.}

\item{sigma}{the standard deviation of the experimental error.}
}
\value{
a data.frame containing the sample size (nreps), the total number of runs of the experiment (ntotal), the residual's degrees of freedom (df2), and the calculated power associated with the effects in the design.
}
\description{
Implementation of the power function for F tests involving one-way and two-way ANOVA using the conservative method based on a single specified difference between any two treatment means, and an standard deviation of the experimental error. The Fpower function has been adapted from the Fpower1 and Fpower2 functions provided in the package daewr.
}
\examples{
# one-way ANOVA:
Fpower(alpha=.05, nlev=3, nreps=2:10, Delta=40, sigma=25)

# two-way ANOVA:
Fpower(alpha=.05, nlev=c(3, 3), nreps=2:10, Delta=40, sigma=25)

# tree-way ANOVA:
Fpower(alpha=.05, nlev=c(3, 3, 3), nreps=2:10, Delta=40, sigma=25)

}
