% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biblioteca.R
\name{gg_daniel}
\alias{gg_daniel}
\title{Daneil probability plot}
\usage{
gg_daniel(tb, alpha = 0.05)
}
\arguments{
\item{tb}{a data frame containing the output of the table2kunrep function}

\item{alpha}{significance level for Lenth's method}
}
\value{
the Daniel probability plot.
}
\description{
ggplot implementation for the Daniel probability plot
}
\examples{
\donttest{
library(planex)
library(tidyverse)

data(filtragem)

filtragem <- filtragem \%>\%
  mutate_at(
    c("temperatura", "pressao", "concentracao", "agitacao"),
    as.factor2k) \%>\%
  rename(
    A = temperatura,
    B = pressao,
    C = concentracao,
    D = agitacao
  )

mod <- lm(filtragem ~ A*B*C*D, data=filtragem)
tb <- table2kunrep(mod)
gg_daniel(tb)
}

}
