% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biblioteca.R
\name{ggresiduals}
\alias{ggresiduals}
\title{Residuals' plots for ANOVA models}
\usage{
ggresiduals(object)
}
\arguments{
\item{model}{an object of the class lm or aov.}
}
\value{
residual plots suitable for residual analysis of data from design experiments.
}
\description{
This function is only suitable to analyze residuals of ANOVA models. It must not be used to analyze residuals from regression or ANCOVA models.
}
\examples{
library(planex)
library(tidyverse)
data(saquinhos)
saquinhos <- mutate(saquinhos,
  concentracao = as.factor(concentracao)
)
mod <- aov(resistencia ~ concentracao, data = saquinhos)
ggresiduals(mod)

}
