% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{impureza}
\alias{impureza}
\title{Impureza de um produto químico}
\format{
Data frame com 15 linhas e 3 variáveis:
\itemize{
  \item temperatura: variável categórica com 3 níveis (fator 1)
  \item pressão: variável categórica com 5 níveis (fator 2)
  \item impureza: impureza do produto químico (variável resposta)
}
}
\description{
Exemplo 5.2: A impureza de um produto químico é afetada por dois fatores, a saber: temperatura e pressão. Um exeprimento com dois fatores fixos cruzados foi conduzido para avaliar o efeito desses dois fatores sobre a impureza do produto químico.
}
\references{
Montgomery, D. C. (2016) Design and Analysis of Experiments, 8ª. Edição, New York: Wiley.
}
\author{
Fábio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{datasets}
