% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{interacao1}
\alias{interacao1}
\title{Entendendo efeitos de interação (caso 1)}
\format{
Data frame com 8 linhas e 3 variáveis:
\itemize{
  \item fator A: variável categórica com 2 níveis
  \item fator B: variável categórica com 2 níveis
  \item resp: variável resposta
}
}
\description{
Exemplo de efeitos de interação em delineamentos com dois fatores fixos cruzados.
}
\references{
Montgomery, D. C. (2016) Design and Analysis of Experiments, 8ª. Edição, New York: Wiley.
}
\author{
Fabio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{datasets}
