% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biblioteca.R
\name{interactionPlot}
\alias{interactionPlot}
\title{Gráfico de interações para delineamentos com 2 a 4 fatores}
\usage{
interactionPlot(modelo)
}
\arguments{
\item{modelo}{modelo ajustado (objeto deve ter classe aov ou lm)}
}
\value{
Retorna o gráfico de interações
}
\description{
Gráficos de interações envolvendo delineamentos com dois a quatro fatores
}
