% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biblioteca.R
\name{interactionTest}
\alias{interactionTest}
\title{Tukey's nonadditivity interaction test.}
\usage{
interactionTest(object)
}
\arguments{
\item{object}{fitted model (an object of class  aov or lm)}
}
\value{
Retorna a tabela ANOVA
}
\description{
Function performs the Tukey's nonadditivity interaction test.

Tukey's nonadditivity interaction test for two-away factorial designs with a single replicate.
}
\examples{
\donttest{
library(planex)
library(tidyverse)
head(impureza)
sapply(impureza, class)
impureza <- mutate(impureza,
                  temperatura = as.factor(temperatura),
                  pressao = as.factor(pressao)
)

mod <- aov(impureza~temperatura+pressao, data=impureza)
interactionTest(mod)
}

}
\references{
Montgomery, D. C. (2016) Design and Analysis of Experiments, 8-th. Edition, New York: Wiley.
}
