% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biblioteca.R
\name{interPlot2}
\alias{interPlot2}
\title{Gráfico de interações para delineamentos com 2 fatores}
\usage{
interPlot2(modelo)
}
\arguments{
\item{modelo}{modelo ajustado (objeto deve ter classe aov ou lm)}
}
\value{
Retorna o gráfico de interações
}
\description{
Gráficos de interações envolvendo delineamentos com dois fatores
}
\examples{
\donttest{
library(planex)
# Exemplo da impureza de um produto químico:
head(impureza)
sapply(impureza, class)
impureza <- mutate(impureza,
                  temperatura = as.factor(temperatura),
                  pressao = as.factor(pressao)
)

mod <- aov(impureza~temperatura+pressao, data=impureza)
interPlot(mod)

# Exemplo do tempo de vida de baterias:
head(baterias)
sapply(baterias, class)
baterias <- mutate(baterias,
                  temperatura = as.factor(temperatura),
                  tipo = as.factor(tipo)
)

mod <- aov(tempo~temperatura+tipo, data=baterias)
interPlot(mod)

}

}
