% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biblioteca.R
\name{interPlot3}
\alias{interPlot3}
\title{Gráfico de interações para delineamentos com 3 fatores}
\usage{
interPlot3(modelo)
}
\arguments{
\item{modelo}{modelo ajustado (objeto deve ter classe aov ou lm)}
}
\value{
Retorna o gráfico de interações
}
\description{
Gráficos de interações envolvendo delineamentos com dois fatores
}
\examples{
\donttest{
library(planex)
# Exemplo das garrafas de refrigerante:
head(refrigerantes)
sapply(refrigerantes, class)
mod1 <- aov(desvio~carbonatacao+pressao+velocidade, data=refrigerantes)
mod2 <- aov(desvio~pressao+carbonatacao+velocidade, data=refrigerantes)
mod3 <- aov(desvio~velocidade+carbonatacao+pressao, data=refrigerantes)
mod4 <- aov(desvio~velocidade+pressao+carbonatacao, data=refrigerantes)

interPlot3(mod1)
interPlot3(mod2)
interPlot3(mod3)
interPlot3(mod4)

}

}
