% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biblioteca.R
\name{interPlot4}
\alias{interPlot4}
\title{Gráfico de interações para delineamentos com 4 fatores}
\usage{
interPlot4(modelo)
}
\arguments{
\item{modelo}{modelo ajustado (objeto deve ter classe aov ou lm)}
}
\value{
Retorna o gráfico de interações
}
\description{
Gráficos de interações envolvendo delineamentos com 4 fatores
}
\examples{
\donttest{
library(planex)
# Exemplo das garrafas de refrigerante:
head(refrigerantes)
sapply(refrigerantes, class)
mod1 <- aov(filtragem ~ temperatura+pressao+concentracao+agitacao, data=filtragem)
mod2 <- aov(filtragem ~ pressao+temperatura+concentracao+agitacao, data=filtragem)
mod3 <- aov(filtragem ~ concentracao+temperatura+pressao+agitacao, data=filtragem)
mod4 <- aov(filtragem ~ agitacao+pressao+temperatura+concentracao, data=filtragem)

interPlot4(mod1)
interPlot4(mod2)
interPlot4(mod3)
interPlot4(mod4)

}

}
