% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biblioteca.R
\name{interTest}
\alias{interTest}
\title{Teste de não aditividade de Tukey}
\usage{
interTest(modelo)
}
\arguments{
\item{modelo}{modelo ajustado (objeto deve ter classe aov ou lm)}
}
\value{
Retorna a tabela ANOVA
}
\description{
Implementação do teste de não aditividade proposto por Tukey (1949) para testar o efeito de interação envolvendo delineamentos com dois fatores fixos cruzados e  uma  única replicação.

Realiza o teste de nao aditividade para delineamentos com dois fatores fixos cruzados.
}
\examples{
\donttest{
library(planex)
head(impureza)
sapply(impureza, class)
impureza <- mutate(impureza,
                  temperatura = as.factor(temperatura),
                  pressao = as.factor(pressao)
)

mod <- aov(impureza~temperatura+pressao, data=impureza)
interTest(mod)
}

}
\references{
Montgomery, D. C. (2016) Design and Analysis of Experiments, 8ª. Edição, New York: Wiley.
}
