% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{medicoes}
\alias{medicoes}
\title{Sistema de medições}
\format{
Data frame com 12 linhas e 3 variáveis:
\itemize{
  \item peca: peça que será submetida à medição (fator 1)
  \item operador: operador operar o equipamento (fator 2)
  \item dimensao: dimensão da peça (variável resposta)
}
}
\description{
Delineamento com dois fatores aleatórios conduzido com o objetivo de avaliar potenciais fontes de variabilidade de um sistema de medições.
}
\references{
Montgomery, D. C. (2016) Design and Analysis of Experiments, 8ª. Edição, New York: Wiley.
}
\author{
Fabio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{datasets}
