% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{moinho}
\alias{moinho}
\title{Homogeneidade da produção de teares}
\format{Data frame com 25 linhas e duas variáveis:
\itemize{
  \item producao: produção do tear em 1b/min. (variável resposta)
  \item tear: variável indicadora dos teares selecionados aleatoriamente da linha de produção (variável categórica com quadro níveis; fator de interesse)
}}
\description{
(Exercício 13.29) Delineamento com um fator aleatório conduzido para avaliar a homogeneidade da produção dos teares de um moinho têxtil.
}
\references{
Montgomery, D. C., Runger, G. C. (2015) Estatística Aplicada e Probabilidade para Engenheiros, 5ª. Edição, Rio de Janeiro: LTC.
}
\author{
Fabio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{datasets}
