% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{montagem}
\alias{montagem}
\title{Tempo de montagem de um componente de tv}
\format{
Data frame com 16 linhas e 5 variáveis:
\itemize{
  \item ordem: ordem de montagem (variável categórica com 4 níveis; bloco 1)
  \item operador: operador selecionado (variável categórica com 4 níveis; bloco 2)
  \item local:  local de trabaho (variável categórica com 4 níveis; bloco 3)
  \item metodo: método de montagem (variável categórica com 4 níveis; fator de interesse principal)
  \item tempo: tempo de montagem da componente
}
}
\description{
Exercício 4.36: Um engenheiro industrial está investigando o efeito de 4 tipos de métodos de montagem sobre o tempo de montagem de um componente de tv. Quatro operadores foram selecionados para o  estudo. Além disso, o engenheiro sabe que que cada processo de montagem gera fadiga (no operador), afetando o tempo de montagem das sucessivas componentes. O engenheiro também suspeita que o local de trabalho dos operadores pode afetar o tempo de montagem. Desta forma, um experimento do tipo quadrado greco-latino foi conduzido, visando controlar as fontes de variação de perturbação. Analise adequadamente os dados desse experimento, apresentando as conclusões práticas.
}
\references{
Montgomery, D. C. (2016) Design and Analysis of Experiments, 8ª. Edição, New York: Wiley.
}
\author{
Fabio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{datasets}
