% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{papel}
\alias{papel}
\title{Efeito da concentração de madeira sobre a força do papel}
\format{
Data frame com 21 linhas e 3 variáveis:
\itemize{
  \item concentracao: concentração de madeira (variável categórica com 7 níveis; fator de interesse principal)
  \item dia: dia em que o experimento foi conduzido (variável categórica com 4 níveis; bloco)
  \item forca: forca do papel (variável resposta)
}
}
\description{
Exercício 4.42: Sete concentrações diferentes de madeira estão sendo estudados para determinar seu efeito sobre a força do papel produzido. No entanto, a planta piloto só pode produzir três corridas do experimento por dia. Como os dias podem afetar a força do papel produzido, um delineamento com blocos incompletos aleatorizados foi conduzido. Analise os dados desse experimento e apresente as conclusões práticas.
}
\references{
Montgomery, D. C. (2016). Design and Analysis of Experiments, 8ª. Edição, New York: Wiley.
}
\author{
Fábio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{datasets}
