% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm.test.R
\name{perm.test.default}
\alias{perm.test.default}
\alias{perm.test.formula}
\alias{perm.test.aov}
\title{Two-sample permutation test for differences in the population means}
\usage{
\method{perm.test}{default}(
  x,
  y,
  nsim = 1000,
  seed = NULL,
  alternative = c("two.sided", "less", "greater"),
  ...
)

\method{perm.test}{formula}(
  formula,
  data,
  nsim = 1000,
  seed = NULL,
  alternative = c("two.sided", "less", "greater"),
  ...
)

\method{perm.test}{aov}(object, nsim = 1000, seed = NULL, ...)
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values.}

\item{y}{a (non-empty) numeric vector of data values.}

\item{nsim}{number of permutations to be simulated}

\item{seed}{seed passed to the random number generator; if NULL (default), then a random seed is used}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less". You can specify just the initial letter.}

\item{...}{further arguments passed to or from other methods.}

\item{formula}{a formula of the form lhs ~ rhs where lhs is a numeric variable giving the data values and rhs either 1 for a one-sample or paired test or a factor with two levels giving the corresponding groups.}

\item{data}{an optional matrix or data frame containing the variables in the formula formula. By default the variables are taken from environment(formula).}

\item{object}{an object of class aov}
}
\value{
perm.test returns an object of class perm.test containing information regarding the permutation test carried out.

perm.test returns an object of class perm.test containing information regarding the permutation test carried out.

perm.test returns an object of class perm.test.aov containing information regarding the permutation test carried out
}
\description{
Two-sample permutation test for differences in the population means

Two-sample permutation test for differences in the population means

Analysis of variance based on permutation tests
}
\details{
The permutation test implemented here is based on the difference of the sample means.

The permutation test implemented here is based on the difference of the sample means.
}
\examples{
\donttest{
library(planex)
library(tidyr)
data(catalisadores2)
t.test  <- t.test(catalisadores2$C1, catalisadores2$C2)
t.test
#perm <- perm.test(catalisadores2$C1, catalisadores2$C2, nsim = 1000, seed = 123)
#perm$pvalor
}

\donttest{
library(planex)
library(tidyr)
data(catalisadores2)
#catalisadores2 <- catalisadores2 \%*\%
#  pivot_longer(
#     cols = 1:2,
#     names_to = "catalisador",
#     values_to = "rendimento"
#   )
# t.test  <- t.test(rendimento ~ catalisador, data=catalisadores2)
# t.test
# perm <- perm.test(rendimento ~ catalisador, data=catalisadores2, nsim = 1000, seed = 123)
# perm$pvalor
}

\donttest{
library(planex)
data(refrigerantes)
mod <- aov(desvio ~ carbonatacao*pressao*velocidade, data=refrigerantes)
summary(mod)
perm <- perm.test(mod)
summary(perm)
}

}
