% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm.test.R
\name{perm.test}
\alias{perm.test}
\title{Generic S3 method cross_time}
\usage{
perm.test(object, ...)
}
\arguments{
\item{object}{an object of the class aov}

\item{...}{further arguments passed to or from other methods}
}
\value{
a list containing the results of a permutation test
}
\description{
Generic S3 method cross_time
}
