% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{pilhas}
\alias{pilhas}
\title{Tempo de vida de pilhas para dispositivos eletrônicos.}
\format{
Data frame com 12 linhas e 3 variáveis:
\itemize{
\item pilha: marca da pilha (A; B) - fator fixo com 2 níveis
\item dispositivo: tipo de dispositivo eletrônico (rádio, câmera e DVD player portátil) - fator fixo com 3 níveis
\item horas: tempo de vida em horas (variável resposta)
}
}
\description{
Um experimento foi conduzido para estudar a vida (em horas) de duas marcas diferentes de baterias em três dispositivos diferentes (rádio, câmera e DVD player portátil). Um delineamento com dois fatores fixos cruzados foi conduzido para investigar os efeitos do tipo de dispositivo e da marca da pilha sobre o tempo de vida das pilhas.
}
\references{
Montgomery, D. C. (2016). Design and Analysis of Experiments, 8ª. Edição, New York: Wiley.
}
\author{
Fábio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{datasets}
