% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planex-package.R
\docType{package}
\name{planex-package}
\alias{planex-package}
\alias{planex}
\title{Pacote planex - Planejamento de Experimentos.}
\references{
Montgomery, D. C. (2016) Design and Analysis of Experiments, 8ª. Edição, New York: Wiley.

Montgomery, D. C., Runger, G. C. (2015) Estatística Aplicada e Probabilidade para Engenheiros, 5ª. Edição, Rio de Janeiro: LTC.
}
