% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biblioteca.R
\name{plotResiduals}
\alias{plotResiduals}
\title{Residual plot checks}
\usage{
plotResiduals(model)
}
\arguments{
\item{model}{an object of the class lm or aov.}
}
\value{
residual plots suitable for residual analysis of data from design experiments.
}
\description{
Residual plot checks
}
\examples{
library(planex)
library(tidyverse)
data(saquinhos)
saquinhos <- mutate(saquinhos,
  concentracao = as.factor(concentracao)
)
mod <- aov(resistencia ~ concentracao, data = saquinhos)
plotResiduals(mod)

}
