% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ztest.R
\name{print.z.test}
\alias{print.z.test}
\title{Print method for z.test objects}
\usage{
\method{print}{z.test}(x, digits = getOption("digits"), prefix = "\\t", ...)
}
\arguments{
\item{x}{x}

\item{digits}{digits}

\item{prefix}{prefix}

\item{...}{description}
}
\description{
Print method for z.test objects
}
