% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm.test.R
\name{print.summary.perm.test.aov}
\alias{print.summary.perm.test.aov}
\title{Print the summary.perm.test.aov output}
\usage{
\method{print}{summary.perm.test.aov}(x, ...)
}
\arguments{
\item{x}{an object of the class perm.test.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
No return value, called for side effects.
}
\description{
Produces a printed summary of an object of the class perm.test.
}
