% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{queima}
\alias{queima}
\title{Acabamento de superfície de peças metálicas}
\format{
Data frame com 18 linhas e três variáveis:
\itemize{
\item densidade: variável resposta
\item posicao: posição da fornalha (fator 1; dois níveis)
\item temperatura: temperatura da fornalha (fator 2; 3 níveis: 800, 825 e 850 °C)
}
}
\description{
Exercício 14.4: Um experimento foi conduzido para determinar se a temperatura (°C) de queima ou a posição da fornalha afetam a densidade de um ânodo de carbono. Análise adequadamente os dados provenientes desse experimento, apresentando as suas conclusões.
}
\references{
Montgomery, D. C., Runger, G. C. (2015) Estatística Aplicada e Probabilidade para Engenheiros, 5ª. Edição, Rio de Janeiro: LTC.
}
\author{
Fábio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{datasets}
