% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{quilometragem}
\alias{quilometragem}
\title{Efeito de aditivos na gasolina sobre a quilometragem}
\format{
Data frame com 20 linhas e 3 variáveis:
\itemize{
  \item aditivo: aditivo utilizado (variável categórica com 5 níveis; fator de interesse principal)
  \item carro: carro utilizado para a realização do experimento (variável categórica com 5 níveis; bloco)
  \item quilometragem: quilometragem observada (variável resposta)
}
}
\description{
Exercício 4.40: Um engenheiro está estudando o desempenho da quilometragem característica de cinco tipos de aditivos à gasolina. No teste de estrada ele deseja usar carros como blocos; no entanto, devido a uma restrição de tempo, ele deve usar um delineamento com blocos incompletos. Analise adequadamente os dados desse experimento e apresente as conclusões práticas.
}
\references{
Montgomery, D. C. (2016) Design and Analysis of Experiments, 8ª. Edição, New York: Wiley.
}
\author{
Fábio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{datasets}
