% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{refrigerantes}
\alias{refrigerantes}
\title{Volume de garrafas de refrigerantes}
\format{
Data frame com 24 linhas e 4 variáveis:
\itemize{
  \item devio: desvio do volume nominal da garrafa (em ml; variável resposta)
  \item carbonatacao: percentual de carbonatação (variável categórica com três níveis; fator de interesse principal)
  \item pressao:  pressão (em libras)  (variável categórica com 2 níveis; fator de interesse principal)
  \item velocidade: velocidade da linha de produção (variável categórica com 2 níveis; fator de interesse principal)
}
}
\description{
Delineamento com 3 fatores fixos cruzados conduzido para avaliar o efeito da carbonatação, da pressão de operação, e a da velocidade da linha de produção sobre o desvio do volume nominal de garrafas de refrigerante.
}
\references{
Montgomery, D. C. (2016). Design and Analysis of Experiments, 8ª. Edição, New York: Wiley.
}
\author{
Fábio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{datasets}
