% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{rendimento}
\alias{rendimento}
\title{Rendimeno de um processo químico}
\format{
Data frame com 25 linhas e 5 variáveis:
\itemize{
\item lote: lote de materia prima utilizado no preparo das formulações (variável categórica com 5 níveis; bloco 1)
\item acido: concentracao de ácido utilizado no processo (variável categórica com 5 níveis; bloco 2)
\item catalisador:  concentração de catalisador utilizado no processo (variável categórica com 5 níveis; bloco 3)
\item tempo: tempo de espera (variável categórica com 5 níveis; fator de interesse principal)
\item rendimento: rendimento do processo químico
}
}
\description{
Exercício 4.35: O rendimento de um processo químico foi medido usando-se 5 lotes de matéria prima, 5 concentrações de ácido, 5 tempos de espera, e 5 concentrações de catalisadores, configurando um delineamento do tipo quadrado greco-latino. Analise adequadamente os dados provenientes desse experimento, apresentando as suas conclusões práticas.
}
\references{
Montgomery, D. C. (2016). Design and Analysis of Experiments, 8ª. Edição, New York: Wiley.
}
\author{
Fábio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{datasets}
