% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{rugosidade}
\alias{rugosidade}
\title{Rugosidade da superfície}
\format{
Data frame com 16 linhas e 4 variáveis:
\itemize{
\item rugosidade: rugosidade (variável resposta)
\item alimentacao: taxa de alimentacao (variável categórica com 2 níveis; fator 1)
\item profundidade: profundade do corte (variável categórica com 2 níveis; fator 2)
\item angulo:  ângulo da ferramenta (variável categórica com 2 níveis; fator 3)
}
}
\description{
Um engenheiro mecânico está estudando a rugosidade da superfície de uma peça usinada. Três fatores são de interesse: taxa de alimentaçao, profundidade do corte e ângulo da ferramenta. O engenheiro decidiu realizar um experinto 2^3 fatorial com duas réplicas completas do experimento
}
\references{
Montgomery, D. C., Runger, G. C. (2015) Estatística Aplicada e Probabilidade para Engenheiros, 5ª. Edição, Rio de Janeiro: LTC.
}
\author{
Fábio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{datasets}
