% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{saquinhos}
\alias{saquinhos}
\title{Resistência de saquinhos de papel}
\format{
Data frame com 24 linhas e duas variáveis:
\itemize{
\item resistencia: resistência dos saquinhos (variável resposta)
\item concentracao: concentração de madeira de lei (variável categórica com quadro níveis; fator de interesse)
}
}
\description{
Exemplo 13.1: Delineamento com um fator fixo conduzido para avaliar o efeito de 4 diferentes concentrações de madeira de lei sobre a resistência de saquinhos de papel.
}
\references{
Montgomery, D. C., Runger, G. C. (2015) Estatística Aplicada e Probabilidade para Engenheiros, 5ª. Edição, Rio de Janeiro: LTC.
}
\author{
Fábio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{datasets}
