% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{semicondutores}
\alias{semicondutores}
\title{Planta de fabricação de semicondutores}
\format{
Data frame com 16 linhas e duas variáveis:
\itemize{
  \item rendimento: variável resposta
  \item comb_trat: combinação de tratamentos.
}
}
\description{
Exemplo 8.2: Um experimento foi realizado em uma planta de fabricação de semicondutores, como um esforço para aumentar o rendimento. Cinto fatores, cada um com dois níveis foram estudados. Os fatores (e níveis) foram A = abertura (pequena, grande), B = tempo de exposição (20% abaixo do normal, 20% acima do normal), C =  tempo de desenvolvimento (30s e 45s), D = dimensão da máscara (pequena, grande), e E = tempo de ataque químico (14.5 min, 15.5 min). Devido ao número elevado de fatores (5) e a questões orçamentárias, uma meia fração do delineamento 2^5 foi corrida. Analise os dados do experimento e tire as suas conclusões.
}
\references{
Montgomery, D. C. (2016) Design and Analysis of Experiments, 8ª. Edição, New York: Wiley.
}
\author{
Fábio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{datasets}
