% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{semicondutores2}
\alias{semicondutores2}
\title{Ataque químico para remoção de silicone em semicondutores}
\format{
Data frame com 10 linhas e duas variáveis:
\itemize{
\item solucao: tipo de solução empregada
\item taxa: taxa de ataque químico (10e-3 polegada/min)
}
}
\description{
Exercício 10.19: Na fabricação de semicondutores, o ataque químico por via úmida é freuentemente usado para remover silicone da parte posterior das pastilhas antes da metalização. A taxa de ataque é uma característica importante nesse processo e é sabido que ela segue uma distribuição normal. Duas soluções para ataque químico foram comparadas, utilizando-se duas amostras aleatórias de 10 pastilhas para cada solução.
}
\references{
Montgomery, D. C., Runger, G. C. (2015) Estatística Aplicada e Probabilidade para Engenheiros, 5ª. Edição, Rio de Janeiro: LTC.
}
\author{
Fábio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{datasets}
