% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm.test.R
\name{summary.perm.test.aov}
\alias{summary.perm.test.aov}
\title{Summary for an perm.test.aov object}
\usage{
\method{summary}{perm.test.aov}(object, ...)
}
\arguments{
\item{object}{the result of a call to summary.perm.test.aov}

\item{...}{further arguments passed to or from other methods.}

\item{conf.level}{the confidence level required.}
}
\value{
an object of the class summary.survstan containing a summary of the fitted model.
}
\description{
Summary for an perm.test.aov object
}
