% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biblioteca.R
\name{table2kunrep}
\alias{table2kunrep}
\title{Summary table for unreplicated/fractional 2^k designs.}
\usage{
table2kunrep(object)
}
\arguments{
\item{object}{fitted model of class aov or lm.}
}
\value{
a data.frame containing the estimated effects, the sum of squares (SS), the sum o squares percentual contribution (SSPC), the effects' relative difference with respect to the overal mean (PRD).
}
\description{
This function is aimed to provide a summary suitable for unreplicated/fractional designs.
}
\examples{
library(planex)
data(semicondutores1rep)
semicondutores1rep <- treatcomb2factors(semicondutores1rep)
fit <- lm(rendimento ~ A*B*C*D*E, data = semicondutores1rep)
table2kunrep(fit)

}
