% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biblioteca.R
\name{table_signs}
\alias{table_signs}
\title{Table of signs for \eqn{2^k} factorial designs}
\usage{
table_signs(k)
}
\arguments{
\item{k}{an integer specifying the number of factors to be considred.}
}
\value{
a data.frame corresponding to the table of signs
}
\description{
This functions provides the table of signs for factorial designs with up to k = 10 factors.
}
\examples{
\donttest{
library(planex)
library(tidyverse)

table_signs(3)

}

}
