% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{teares}
\alias{teares}
\title{Homogeneidade da produção de teares}
\format{
Data frame com 16 linhas e duas variáveis:
\itemize{
  \item resistencia: resistência dos tecidos (variável resposta)
  \item tear: variável indicadora dos teares selecionados aleatoriamente da linha de produção (variável categórica com quadro níveis; fator de interesse)
}
}
\description{
Delineamento com um fator aleatório conduzido para avaliar a homogeneidade da resistência de tecidos produzidos por diferentes teares de uma fábrica.
}
\references{
Montgomery, D. C., Runger, G. C. (2015) Estatística Aplicada e Probabilidade para Engenheiros, 5ª. Edição, Rio de Janeiro: LTC.
}
\author{
Fábio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{datasets}
