% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{tecidos}
\alias{tecidos}
\title{Resistência de tecidos}
\format{
Data frame com 20 linhas e 3 variáveis:
\itemize{
\item resistencia: resistência do tecido (variável resposta)
\item produto: tipo de produto químico (variável categórica com quadro níveis; fator de interesse)
\item amostra: amostra de tecido utilizada no experimento (bloco; fator de interesse secundário)
}
}
\description{
Exemplo 13.5: Delineamento em esquema de blocos completos aleatorizados e um fator fixo conduzido para avaliar o efeito de 4 diferentes produtos químicos sobre a resistência de tecidos.
}
\references{
Montgomery, D. C., Runger, G. C. (2015) Estatística Aplicada e Probabilidade para Engenheiros, 5ª. Edição, Rio de Janeiro: LTC.
}
\author{
Fábio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{datasets}
