% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{televisores}
\alias{televisores}
\title{Tratamentos de retardantes de fogo}
\format{
Data frame com 18 linhas e três variáveis:
\itemize{
\item corrente: variável resposta (em microampères)
\item vidro: tipo de vidro (fator 1; 2 níveis)
\item fosforo: tipo de fósforo (fator 2; 3 níveis)
}
}
\description{
Exercício 14.1: Um engenheiro conduziu um experimento para investigar o feito de dois fatores (tipo de vidro e tipo de fósforo) sobre o brilho de um tubo de televisão. A variável resposta medida é a corrente (em microampères) necessária para a obtenção de um nível especificado de brilho. Análise adequadamente os dados provenientes desse experimento, apresentando as suas conclusões.
}
\references{
Montgomery, D. C., Runger, G. C. (2015) Estatística Aplicada e Probabilidade para Engenheiros, 5ª. Edição, Rio de Janeiro: LTC.
}
\author{
Fábio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{datasets}
