% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biblioteca.R
\name{testResiduals}
\alias{testResiduals}
\title{Residual's formal tests for ANOVA models}
\usage{
testResiduals(
  model,
  normality.test = c("sw", "ad"),
  var.test = c("bartlett", "levene")
)
}
\arguments{
\item{model}{an object of the class lm or aov.}

\item{normality.test}{type of normality test; currently Shapiro-Wilk (SW) and Anderson-Darling (AD) tests are available}

\item{var.test}{type of homogeneity test of variances; currently Bartlett and Levene tests are available}
}
\value{
residuals' normality test and residuals' homoneity test of variances.
}
\description{
This function is only suitable to analyze residuals of ANOVA models. It must not be used to analyze residuals from regression or ANCOVA models.
}
\examples{
library(planex)
library(tidyverse)
baterias <- mutate(baterias,
  temperatura = as.factor(temperatura),
  tipo = as.factor(tipo)
)

mod <- aov(tempo ~ temperatura*tipo, data=baterias)
testResiduals(mod)

}
