% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{turbinas2}
\alias{turbinas2}
\title{Taxa de queima de turbinas de aviões}
\format{
Data frame com 36 linhas e 3 variáveis:
\itemize{
  \item queima: taxa de queira (variável resposta).
  \item processo: processo de produção (fator fixo com 3 níveis).
  \item lote: lote de matéria prima (fator aleatório com 4 níveis).
}
}
\description{
Exercício 14.1: Um fabricante de turbinas de aviões está interessado em estudar a taxa de queima de turbinas provenientes de três processos de produção. Quatro lotes de turbinas são escolhidas ao acaso a partir da saída de cada um dos três processos de interesse, e três determinações da velocidade de combustão são feitas em cada lote. Pede-se: análise adequadamente os dados provenientes desse experimento.
}
\references{
Montgomery, D. C. (2016). Design and Analysis of Experiments, 8ª. Edição, New York: Wiley.
}
\author{
Fábio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{datasets}
