% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{vigas}
\alias{vigas}
\title{Resistência de vigas de aço}
\format{
Data frame com 18 linhas e duas variáveis:
\itemize{
  \item metodo: método empregado(K ou L)
  \item previsao: previsão de resistência ao cisalhamento das vigas de aço (variável resposta)
}
}
\description{
Exemplo 10.10: Um pesquisador tem interesse em comparar dois métodos, a saber Karlruhe (K) e Lehigh (L), para prever a resistência ao cisalhamento entre traves planas metálicas. Os dois métodos são aplicados a 9 traves especı́ficas.
}
\references{
Montgomery, D. C., Runger, G. C. (2015) Estatística Aplicada e Probabilidade para Engenheiros, 5ª. Edição, Rio de Janeiro: LTC.
}
\author{
Fábio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{datasets}
