% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{vinhos}
\alias{vinhos}
\title{Competição de vinhos}
\format{
Data frame com 24 linhas e 3 variáveis:
\itemize{
  \item score: pontuação atribuída ao vinho.
  \item juiz: degustador do vinho (variável categórica com 6 níveis).
  \item vinho: vinhos participantes da competição (variável categória com 4 níveis).
}
}
\description{
Exemplo: Em uma competição de julgamento de vinhos, quatro vinhos Chardonnay da mesma safra foram julgados por seis juízes, selecionados aleatoriamente de uma lista de sommeliers renomados. Cada juiz provou os vinhos de forma cega, ou seja, sem conhecer suas identidades. A ordem da apresentação do vinho foi aleatorizada de forma independente para cada juiz. Cada vinho foi avaliado em uma escala de 40 pontos; quanto maior a pontuação, maior é a excelência do vinho.
}
\references{
Kutner, M. H.,  Nachtsheim, C. J., Neter, J., Li, W. (2005). Applied Linear Statistical Models, 5th Edition, New York: McGraw-Hill.
}
\author{
Fábio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{datasets}
