% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biblioteca.R
\name{yates2factors}
\alias{yates2factors}
\title{Yates order to canonical format}
\usage{
yates2factors(data)
}
\arguments{
\item{data}{n x 2 data frame containing the experimental data.}
}
\value{
the updated data with factors in the canonical format
}
\description{
This function translates treatment combinations organized according to Yates order into factors following the $2^k$ factorial canonical format.
}
\examples{
library(planex)
data(refrigerantes2k)
head(refrigerantes2k)
refrigerantes2k <- yates2factors(refrigerantes2k)
refrigerantes2k

}
