% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ztest.R
\name{z.test.default}
\alias{z.test.default}
\title{z.test default implementation}
\usage{
\method{z.test}{default}(
  x,
  y = NULL,
  alternative = c("two.sided", "less", "greater"),
  mu = 0,
  sigma,
  conf.level = 0.95,
  data = NULL
)
}
\arguments{
\item{x}{x}

\item{y}{y}

\item{alternative}{alternative}

\item{mu}{mu}

\item{sigma}{sigma}

\item{conf.level}{conf.level}

\item{data}{data}
}
\description{
z.test default implementation
}
