% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ztest.R
\name{z.test.formula}
\alias{z.test.formula}
\title{z.test formula implementation}
\usage{
\method{z.test}{formula}(
  formula,
  alternative = c("two.sided", "less", "greater"),
  mu = 0,
  sigma,
  conf.level = 0.95,
  data = NULL
)
}
\arguments{
\item{formula}{formula}

\item{alternative}{alternative}

\item{mu}{mu}

\item{sigma}{sigma}

\item{conf.level}{conf.level}

\item{data}{data}
}
\description{
z.test formula implementation
}
