% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ztest.R
\name{z.test}
\alias{z.test}
\title{Generic S3 method z.test}
\usage{
z.test(object, ...)
}
\arguments{
\item{object}{a fitted model object}

\item{...}{further arguments passed to or from other methods.}
}
\value{
the z-test
}
\description{
Generic S3 method z.test
}
