% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{zarcao}
\alias{zarcao}
\title{Zarcao de tintas para aviões}
\format{
Data frame com 18 linhas e três variáveis:
\itemize{
  \item adesao: adesao da tinta (variável resposta)
  \item tipo: tipo de zarcao (variável categórica com 3 níveis; fator 1)
  \item metodo: métido de aplicação do zarcao (variável categórica com 2 níveis; fator 2)
}
}
\description{
Delineamento com um fator fixo conduzido para avaliar o efeito de 4 diferentes concentrações de madeira de lei sobre a resistência de saquinhos de papel.
}
\references{
Montgomery, D. C., Runger, G. C. (2015) Estatística Aplicada e Probabilidade para Engenheiros, 5ª. Edição, Rio de Janeiro: LTC.
}
\author{
Fábio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{datasets}
