% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{acabamento}
\alias{acabamento}
\title{Acabamento de superfície de peças metálicas}
\format{
Data frame com 18 linhas e três variáveis:
\itemize{
  \item acabamento: variável resposta
  \item tinta: tipo de tinta utilzada (fator 1; dois níveis)
  \item tempo: tempo de secagem (fator 2; 3 níveis: 20min, 25min e 30min)
}
}
\description{
Exercício 14.2: Um engenheiro suspeita que o acabamento de uma superfícies de peças metálicas seja influenciado pelo tipo de tinta usada e pelo tempo de secagem. Ele selecionou 3 tempos de secagem - 20, 25, e 30 minutos - e usou dois tipos de tinta. Três peças são testadas com cada combinação de tipo de tinta e tempo de secagem. Realize a análise dos dados e tire as suas conclusões.
}
\references{
Montgomery, D. C., Runger, G. C. (2015) Estatística Aplicada e Probabilidade para Engenheiros, 5ª. Edição, Rio de Janeiro: LTC.
}
\author{
Fábio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{datasets}
