% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biblioteca.R
\name{as.factor2k}
\alias{as.factor2k}
\title{Codificação de variáveis para delineamentos $2^k$ fatoriais}
\usage{
as.factor2k(x)
}
\arguments{
\item{x}{fator que se deseja codificar.}
}
\value{
o fator codificado em termos dos níveis baixo (-1) e alto (+1).
}
\description{
Função para codificar os níveis baixos e altos de fatores de interesse em um delinemanto $2^K$ tal que -1 representa o nível baixo e 1 representa o nível alto do fator.
}
\examples{
library(planex)
library(tidyverse)
data(rugosidade)
head(rugosidade)
rugosidade <- mutate(rugosidade,
  alimentacao = as.factor2k(alimentacao),
  profundidade = as.factor2k(profundidade),
  angulo = as.factor2k(angulo)
)
rugosidade


}
