% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{semicondutores}
\alias{semicondutores}
\title{Semicondutores}
\format{
Data frame com 32 linhas e duas variáveis:
\itemize{
  \item rendimento:
  \item comb_trat: combinação de tratamentos na ordenação de Yates.
}
}
\description{
Exercício 14.17:
}
\references{
Montgomery, D. C., Runger, G. C. (2015) Estatística Aplicada e Probabilidade para Engenheiros, 5ª. Edição, Rio de Janeiro: LTC.
}
\author{
Fábio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{datasets}
