% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biblioteca.R
\name{yates2factors}
\alias{yates2factors}
\alias{comb2factors}
\title{Treatment combinations to canonical format}
\usage{
yates2factors(data)
}
\arguments{
\item{data}{n x 2 data frame containing the experimental data.}
}
\value{
the updated data with factors in the canonical format
}
\description{
This function uses treatment combinations in standard (Yates') order to generate factors canonical format for factorial designs with only two levels.
}
\examples{
library(planex)
data(refrigerantes2k)
head(refrigerantes2k)
refrigerantes2k <- comb2factors(refrigerantes2k)
refrigerantes2k

}
