% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{zarcao}
\alias{zarcao}
\title{Zarcões de tintas para aviões}
\format{
Data frame com 18 linhas e três variáveis:
\itemize{
  \item adesao: adesao da tinta (variável resposta)
  \item tipo: tipo de zarcao (variável categórica com 3 níveis; fator 1)
  \item metodo: método de aplicação do zarcao (variável categórica com 2 níveis; fator 2)
}
}
\description{
Exemplo 14.1: Zarcões de tinta para avião são aplicados em superfícies de alumínio, por meio de dois métodos: imersão e aspersão. A finalidade do zarcão é melhorar a adesão da tinta, podendo ser aplicado em algumas peças usando qualquer método. O grupo de engenharia de processos responsável por essa operação está interessado em saber se três diferentes tipos de zarcões diferem quanto as suas propriedades de adesão. Um delineamento fatorial foi realizado para investigar o efeito do tipo de zarcão e do método de aplicação na adesão da tinta. Três espécimes foram pintados com cada um dos zarcões, usando cada um dos métodos de adesão. Uma camada de tinta foi aplicada e a força de adesão foi medida.
}
\references{
Montgomery, D. C., Runger, G. C. (2015) Estatística Aplicada e Probabilidade para Engenheiros, 5ª. Edição, Rio de Janeiro: LTC.
}
\author{
Fábio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{datasets}
